Enable version command again
